[Refactor] Split the selective checkpointing contract below the model layer
The contract names callables that live in `xtuner/v1/module`, and the model
configs that select units live in `xtuner/v1/model`. Keeping the vocabulary
under either package would make the two import each other.

Split rather than move: the vocabulary and the unit marker go to
`xtuner/v1/utils`, below both packages, while the policy and the checkpoint
wrapping stay at the model layer, where knowing `nn.Module` and
`CheckpointWrapper` is legitimate. `xtuner.v1.model.utils` keeps exporting the
same names, so no import site outside these files changes.

The test that guards this imports a decoder layer in a clean interpreter: in
process the model layer is already loaded, so a contract that reached back
into `model/` would still import fine and the cycle would only appear later.
