+++ xtuner/v1/model/utils/__init__.py
@@ -1,31 +1,24 @@
-from .checkpointing import apply_gradient_checkpointing
-from .misc import ModelForwardExtraLogInfo, module_dict_repr
-from .selective_checkpointing import (
+from xtuner.v1.utils.selective_checkpointing import (
     KeptCallables,
     KeptOps,
-    RecomputeTarget,
     RecomputeTargetMap,
     RecomputeUnit,
-    active_recompute_unit,
-    apply_selective_checkpointing,
-    in_recompute_unit,
-    recompute_unit,
-    resolve_kept_ops,
 )
+
+from .checkpointing import apply_gradient_checkpointing
+from .misc import ModelForwardExtraLogInfo, module_dict_repr
+from .selective_checkpointing import apply_selective_checkpointing, in_recompute_unit, resolve_kept_ops
 
 
 __all__ = [
     "apply_gradient_checkpointing",
     "apply_selective_checkpointing",
-    "in_recompute_unit",
-    "resolve_kept_ops",
     "module_dict_repr",
     "ModelForwardExtraLogInfo",
     "KeptCallables",
     "KeptOps",
-    "RecomputeTarget",
     "RecomputeTargetMap",
     "RecomputeUnit",
-    "active_recompute_unit",
-    "recompute_unit",
+    "in_recompute_unit",
+    "resolve_kept_ops",
 ]
